Add files via upload
Added two additional bits of information: "object" and "special".

"Object" is the memory address of the mirror in the object table. A value of -1 means it doesn't have a door sprite (e.g a wind tunnel).
"Special" used for the feature that changes mirrors appearance based on where they go. 1 = to a switch room, and 2 = a boss chamber. A value of -1 is given to a entrance that does not have a door object to edit.