Add files via upload
Added a feature to change the mirror sprites depending on where they lead. If a one-way leads to a boss mirror, it will have a sparkle effect added to it. If it leads to a switch room, it will instead be the fancy "world mirror" sprite.

This also changes all the "world mirror" objects into regular doors.